move MeasurementOverlay.js to /components
